--- CMA_1/Header1.h
@@ -1,1 +1,0 @@
-#pragma once
