+++ CMA_1/__vm/.CMA_1.vsarduino.h
@@ -111,8 +111,6 @@ typedef long pthread_cond_t;
 #include "MQTT.ino"
 #include "SDCARD.ino"
 #include "SaveData.ino"
-#include "TASK_Can.ino"
-#include "TASK_RFID.ino"
 #include "TASK_display.ino"
 #include "Task_CheckData.ino"
 #include "UART_Inter.ino"
